requirements.txt
django
djangorestframework
django-filters
drf-spectacular
drf-spectacular-sidecar
urllib
pymongo
Pillow
